pyproject.toml
[project]
name = "tosk-bot"
version = "0.1.0"
description = "Studying containers"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "3.14"
dependencies = [
    "python-telegram-bot (>=22.5,<23.0)",
    "pydantic (>=2.12.3,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)"
]
packages = [
    {include = "bot"}
]

[dependency-groups]
dev = [
    "pre-commit (>=4.3.0,<5.0.0)",
    "black (>=25.9.0,<26.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.pytest.ini_options]
pythonpath = "."
